+++ NeuroEvolution/genome.h
@@ -33,11 +33,11 @@ struct ne_genome
         return nodes.data() + input_size;
     }
     
-    inline size_t size() const {
+    inline uint64 size() const {
         return nodes.size();
     }
     
-    inline size_t complexity() const {
+    inline uint64 complexity() const {
         return genes.size();
     }
     
@@ -46,11 +46,11 @@ struct ne_genome
     
     void _destory_genes();
     
-    size_t create_node();
+    uint64 create_node();
     
-    void reset(size_t inputs, size_t outputs);
+    void reset(uint64 inputs, uint64 outputs);
     
-    void initialize(ne_innov_map* map, size_t* innov);
+    void initialize(ne_innov_map* map, uint64* innov);
     
     void insert(ne_gene* gene);
     
@@ -62,21 +62,21 @@ struct ne_genome
         
     void mutate_weights(const ne_params& params);
     
-    void mutate_topology_add_node(ne_innov_map* map, size_t* innov, const ne_params& params);
+    void mutate_topology_add_node(ne_innov_map* map, uint64* innov, const ne_params& params);
     
-    void mutate_topology_add_gene(ne_innov_map* map, size_t* innov, const ne_params& params);
+    void mutate_topology_add_gene(ne_innov_map* map, uint64* innov, const ne_params& params);
     
-    void mutate_toggle_gene_enable(size_t times);
+    void mutate_toggle_gene_enable(uint64 times);
     
-    double fitness;
-    double adjusted_fitness;
+    float64 fitness;
+    float64 adjusted_fitness;
         
     ne_innov_set set;
     
-    size_t input_size;
-    size_t output_size;
+    uint64 input_size;
+    uint64 output_size;
     
-    size_t activations;
+    uint64 activations;
     
     std::vector<ne_node> nodes;
     std::vector<ne_gene*> genes;
@@ -91,9 +91,9 @@ inline bool ne_genome_adjusted_sort(const ne_genome* a, const ne_genome* b) {
 }
 
 void ne_crossover(const ne_genome* A, const ne_genome* B, ne_genome* C, const ne_params& params);
-double ne_distance(const ne_genome* A, const ne_genome* B, const ne_params& params);
+float64 ne_distance(const ne_genome* A, const ne_genome* B, const ne_params& params);
 
-inline bool ne_mutate(ne_genome* genome, const ne_params& params, ne_innov_map* map, size_t* innov) {
+inline bool ne_mutate(ne_genome* genome, const ne_params& params, ne_innov_map* map, uint64* innov) {
     if(ne_random() < params.new_node_prob) {
         genome->mutate_topology_add_node(map, innov, params);
         return true;
